Add more test coverage for Visual studio with the current model
